[#1375] Revise public API
move LicensesResource and Factory to internal package
