+++ bundles/org.eclipse.passage.lic.licenses.model/META-INF/MANIFEST.MF
@@ -17,6 +17,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="0.0.0";x-installation:=
 Export-Package: org.eclipse.passage.lic.internal.licenses.convert;x-friends:="org.eclipse.passage.lic.hc",
  org.eclipse.passage.lic.internal.licenses.model;x-friends:="org.eclipse.passage.lic.licenses.migration,org.eclipse.passage.loc.dashboard.ui,org.eclipse.passage.loc.licenses.core",
  org.eclipse.passage.lic.internal.licenses.model.signature;x-internal:=true,
+ org.eclipse.passage.lic.internal.licenses.model.util;x-internal:=true,
  org.eclipse.passage.lic.licenses.model.api,
  org.eclipse.passage.lic.licenses.model.described;x-internal:=true,
  org.eclipse.passage.lic.licenses.model.impl;x-internal:=true,
