+++ bundles/org.eclipse.passage.loc.licenses.core/src/org/eclipse/passage/loc/internal/licenses/core/LicenseAgreements.java
@@ -87,6 +87,7 @@ final class LicenseAgreements {
 		}
 	}
 
+	@SuppressWarnings("restriction")
 	private Hashes hashes() throws LicensingException {
 		Optional<HashesRegistry> service = new OperatorGearAware().withGear(gear -> Optional.of(gear.hashes()));
 		if (!service.isPresent()) {
